🔨 refactor(engine.py): rename capital_percentage to token0_capital and change type to float 🔨 refactor(TokenManagement.py): change type of current_price, range_low, range_high, and total_token0_amount to float 🔨 refactor(uniswap.py): change return type of remove_liquidity_and_collect_fees function to Tuple[TxReceipt, TxReceipt, TxReceipt] The changes made are to improve the readability and consistency of the code. The variable capital_percentage in engine.py has been renamed to token0_capital to better reflect its purpose. Its type has also been changed to float to allow for more precise calculations. In TokenManagement.py, the types of current_price, range_low, range_high, and total_token0_amount have been changed to float to allow for more precise calculations. In uniswap.py, the return type of the remove_liquidity_and_collect_fees function has been changed to Tuple[TxReceipt,
🔥 chore(web_manager.py): remove unused InsufficientFunds exception
🚚 refactor(web_manager.py): rename capital_percentage to token0_capital
🚀 feat(web_manager.py): add logging of pool info on initialization
🐛 fix(web_manager.py): calculate price correctly in get_price_weth_per_usdc
The InsufficientFunds exception is no longer used and has been removed. The capital_percentage variable has been renamed to token0_capital to better reflect its purpose. Pool information is now logged on initialization to provide more context. The get_price_weth_per_usdc function has been fixed to calculate the price correctly.

🔨 refactor(web_manager.py): refactor swap_amounts method
✨ feat(web_manager.py): add amount calculation and logging to open_position method
🔨 refactor(web_manager.py): refactor close_position method
The swap_amounts method has been refactored to include the calculation of the required amounts and the swap logic. The open_position method now includes the calculation of the required amounts and logs the current price and existing/required amounts for each token. The close_position method has been refactored to include the removal of liquidity, collection of fees, and burning of LP tokens.

🔧 chore(test_api.py): rename capital_percentage to token0_capital
The variable name capital_percentage was not descriptive enough and could be confused with other types of capital. The new name token0_capital is more specific and indicates that it is the capital for token0.

+++ package/uniswap_math/TokenManagement.py
@@ -183,10 +183,10 @@ class TokenManager:
 
     def calculate_amounts(
         self,
-        current_price: int,
-        range_low: int,
-        range_high: int,
-        total_token0_amount: int,
+        current_price: float,
+        range_low: float,
+        range_high: float,
+        total_token0_amount: float,
     ) -> Tuple[int, int]:
         """Calculates the amounts of token0 and token1.
         Requires to use prices in eth format e.g.: 1800 for $1800
